refactor: 반환형 long으로 수정

+++ src/main/java/com/prism/statistics/application/webhook/LabelRemovedService.java
@@ -40,7 +40,7 @@ public class LabelRemovedService {
         Long pullRequestId = pullRequest.getId();
         LocalDateTime unlabeledAt = toLocalDateTime(request.unlabeledAt());
 
-        int deleted = prLabelRepository.deleteLabel(pullRequestId, labelName);
+        long deleted = prLabelRepository.deleteLabel(pullRequestId, labelName);
         if (deleted == 0) {
             return;
         }
